Extract tests for contains to own file
